+++ ADMF/internal/scriptblocks/scriptblocks.ps1
@@ -12,10 +12,7 @@ Set-PSFScriptblock -Name 'ADMF.ScriptBlockName' -Scriptblock {
 #>
 Set-PSFScriptblock -Name 'ADMF.Validate.Type.Gpo' -Scriptblock {
     foreach ($item in $_) {
-        if (
-			-not ($item -is [Microsoft.GroupPolicy.Gpo]) -or
-			($item.PSObject.TypeNames -contains 'Microsoft.GroupPolicy.Gpo')
-		) { return $false }
+        if ($item.PSObject.TypeNames -notcontains 'Microsoft.GroupPolicy.Gpo') { return $false }
     }
     $true
 }